requirements.txt
streamlit
requests
pillow
duckduckgo-search
